chore: move atoms into folders
