+++ stories/atoms/image/BaseImage.stories.js
@@ -7,7 +7,7 @@ export default {
 
 // create multiple src images to match the srcset sizes
 // just using one size for example
-import src from "../assets/lionel-richie.jpg";
+import src from "../../assets/lionel-richie.jpg";
 const srcset = `${src} 300w, ${src} 600w, ${src} 900w`;
 const alt = "lionel richie";
 const sizes = "(max-width: 500px) 300px, (max-width: 800px) 600px, 100vw";
