+++ stories/atoms/image/BaseImage.vue
