+++ stories/organisms/contact-form/ContactForm.vue
@@ -18,7 +18,7 @@
 
 <script>
 import FormInput from "../../molecules/input-group/InputGroup.vue";
-import Button from "../../atoms/Button.vue";
+import Button from "../../atoms/button/Button.vue";
 import Toast from "../../molecules/toast/Toast.vue";
 export default {
   components: {
