+++ stories/organisms/email-subscribe/EmailSubscribe.stories.js
@@ -2,7 +2,7 @@ import EmailSubscribe from "./EmailSubscribe.vue";
 // Instead of importing the FormInput, we import its stories
 // https://storybook.js.org/docs/react/workflows/stories-for-multiple-components
 import FormInput from "../../molecules/input-group/InputGroup.stories";
-import Button from "../../atoms/Button.stories";
+import Button from "../../atoms/button/Button.stories";
 
 export default {
   title: "organisms/Email Subscribe",
