+++ stories/organisms/email-subscribe/EmailSubscribe.vue
@@ -20,7 +20,7 @@
 
 <script>
 import FormInput from "../../molecules/input-group/InputGroup.vue";
-import Button from "../../atoms/Button.vue";
+import Button from "../../atoms/button/Button.vue";
 export default {
   components: {
     FormInput,
